# backend/api/openapi.yaml
openapi: 3.0.3
info:
  title: Barbossa API
  description: |
    Family music library management API.

    ## Authentication
    All endpoints except `/api/auth/login` require a valid JWT token.
    Pass token in Authorization header: `Bearer <token>`

    ## User Roles
    - **Regular**: Can browse, heart, download, export own library
    - **Admin**: Can manage users, delete from disk, upload to TorrentLeech
  version: 0.1.9
  contact:
    name: Barbossa
  license:
    name: Private
servers:
  - url: http://localhost:8080
    description: Local development
  - url: http://barbossa:8080
    description: Docker internal
...
paths:
...
  /api/artists/{id}:
    get:
      tags: [artists]
      summary: Get artist details
      operationId: getArtist
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ArtistId'
      responses:
        '200':
          description: Artist details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Artist'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/albums/{id}:
    get:
      tags: [albums]
      summary: Get album details
      operationId: getAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Album'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [albums, admin]
      summary: Delete album from disk (admin only)
      operationId: deleteAlbum
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album deleted
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
  /api/albums/{id}/artwork:
    get:
      tags: [albums]
      summary: Get album artwork
      operationId: getAlbumArtwork
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      responses:
        '200':
          description: Album artwork
          content:
            image/jpeg:
              schema:
                type: string
                format: binary
            image/png:
              schema:
                type: string
                format: binary
    put:
      tags: [albums]
      summary: Update album artwork
      operationId: updateAlbumArtwork
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/AlbumId'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                artwork:
                  type: string
                  format: binary
      responses:
        '200':
          description: Artwork updated
        '400':
          $ref: '#/components/responses/BadRequest'

  # ==========================================================================
  # TRACKS
  # ==========================================================================
  /api/tracks/{id}:
    get:
      tags: [tracks]
      summary: Get track details
      operationId: getTrack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Track details with quality info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
  /api/tracks/{id}/metadata:
    put:
      tags: [tracks]
      summary: Update track metadata
      operationId: updateTrackMetadata
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TrackMetadataUpdate'
      responses:
        '200':
          description: Metadata updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Track'
  /api/tracks/{id}/stream:
    get:
      tags: [tracks]
      summary: Stream track for preview
      operationId: streamTrack
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/TrackId'
      responses:
        '200':
          description: Audio stream
          content:
            audio/flac:
              schema:
                type: string
                format: binary
            audio/mpeg:
              schema:
                type: string
                format: binary

  # ==========================================================================
  # SEARCH
  # ==========================================================================
...
  /api/settings:
    get:
      tags: [settings]
      summary: Get all settings
      operationId: getSettings
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Settings
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
    put:
      tags: [settings, admin]
      summary: Update settings (admin only)
      operationId: updateSettings
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        '200':
          description: Settings updated

  # ==========================================================================
  # HEALTH
  # ==========================================================================
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    ArtistId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    AlbumId:
      name: id
      in: path
      required: true
      schema:
        type: integer
    TrackId:
      name: id
      in: path
      required: true
      schema:
        type: integer
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    Forbidden:
      description: Forbidden (admin only)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  schemas:
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        details:
          type: object
...
    Artist:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        album_count:
          type: integer
        artwork_path:
          type: string
...
    Album:
      type: object
      properties:
        id:
          type: integer
        artist_id:
          type: integer
        artist_name:
          type: string
        title:
          type: string
        year:
          type: integer
        artwork_path:
          type: string
        total_tracks:
          type: integer
        available_tracks:
          type: integer
        source:
          type: string
        is_hearted:
          type: boolean
...
    Track:
      type: object
      properties:
        id:
          type: integer
        album_id:
          type: integer
        title:
          type: string
        track_number:
          type: integer
        disc_number:
          type: integer
        duration:
          type: integer
        format:
          type: string
        sample_rate:
          type: integer
        bit_depth:
          type: integer
        is_lossy:
          type: boolean
        source:
          type: string
        source_quality:
          type: string
        is_hearted:
          type: boolean
...
    TrackMetadataUpdate:
      type: object
      properties:
        title:
          type: string
        artist:
          type: string
        album:
          type: string
        track_number:
          type: integer
        year:
          type: integer
        genre:
          type: string
...